cockpit: add balance total
