+++ packages/cockpit/src/balance/mocks/totalExpectedResult.json
@@ -1,0 +1,5 @@
+{
+  "net": 76451657,
+  "outcoming": 88996335,
+  "outgoing": 12544678
+}
