+++ packages/cockpit/src/balance/mocks/totalFromRequestsMock.json
@@ -1,0 +1,122 @@
+[
+  {
+    "date": "2018-03-13T03:00:00.000Z",
+    "available": {
+      "amount": {
+        "balance": 523577340,
+        "in": 30526017,
+        "out": -1640150
+      },
+      "fee": {
+        "balance": 462441404,
+        "in": 7946652,
+        "out": -145946
+      }
+    }
+  },
+  {
+    "date": "2018-03-14T03:00:00.000Z",
+    "available": {
+      "amount": {
+        "balance": 536419728,
+        "in": 14888180,
+        "out": -2045792
+      },
+      "fee": {
+        "balance": 464911449,
+        "in": 2596798,
+        "out": -126753
+      }
+    }
+  },
+  {
+    "date": "2018-03-15T03:00:00.000Z",
+    "available": {
+      "amount": {
+        "balance": 546164504,
+        "in": 12185375,
+        "out": -2440599
+      },
+      "fee": {
+        "balance": 467275452,
+        "in": 2487257,
+        "out": -123254
+      }
+    }
+  },
+  {
+    "date": "2018-03-16T03:00:00.000Z",
+    "available": {
+      "amount": {
+        "balance": 556412012,
+        "in": 12350369,
+        "out": -2102861
+      },
+      "fee": {
+        "balance": 469670943,
+        "in": 2490177,
+        "out": -94686
+      }
+    }
+  },
+  {
+    "date": "2018-03-17T03:00:00.000Z",
+    "available": {
+      "amount": {
+        "balance": 556084536,
+        "in": 61126,
+        "out": -388602
+      },
+      "fee": {
+        "balance": 469661411,
+        "in": 16518,
+        "out": -26050
+      }
+    }
+  },
+  {
+    "date": "2018-03-18T03:00:00.000Z",
+    "available": {
+      "amount": {
+        "balance": 555679691,
+        "in": 784189,
+        "out": -1189034
+      },
+      "fee": {
+        "balance": 469656494,
+        "in": 4277,
+        "out": -9194
+      }
+    }
+  },
+  {
+    "date": "2018-03-19T03:00:00.000Z",
+    "available": {
+      "amount": {
+        "balance": 566729182,
+        "in": 13052383,
+        "out": -2002892
+      },
+      "fee": {
+        "balance": 472217428,
+        "in": 2786013,
+        "out": -225079
+      }
+    }
+  },
+  {
+    "date": "2018-03-20T03:00:00.000Z",
+    "available": {
+      "amount": {
+        "balance": 596510418,
+        "in": 31473545,
+        "out": -1692309
+      },
+      "fee": {
+        "balance": 480007986,
+        "in": 7997157,
+        "out": -206599
+      }
+    }
+  }
+]
